requirements.txt
python-binance>=1.0.19
websocket-client>=1.5.0
aiohttp>=3.8.0
discord.py>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0